One lightning synth
